Add LatentApproxDistance class
